# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HR Workforce Management Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  companyName: string;
  phone: string;
  email: string;
  password: string;
  adminName: string;
}

export interface AdminUser {
  id: number;
  companyId: number;
  name: string;
  email: string;
  role: string;
}

export interface Company {
  id: number;
  name: string;
  phone: string;
  email: string;
  logo?: string | null;
  subscriptionPlan: string;
  createdAt: string;
}

export interface LoginResponse {
  success: boolean;
  user: AdminUser;
  company: Company;
}

export interface MeResponse {
  user: AdminUser;
  company: Company;
}

export interface UpdateCompanyRequest {
  name?: string;
  phone?: string;
  logo?: string;
}

export interface CompanyStats {
  totalEmployees: number;
  presentToday: number;
  absentToday: number;
  lateToday: number;
  totalDevices: number;
  attendanceRate: number;
}

export type EmployeeSalaryType =
  (typeof EmployeeSalaryType)[keyof typeof EmployeeSalaryType];

export const EmployeeSalaryType = {
  hourly: "hourly",
  monthly: "monthly",
} as const;

export interface Employee {
  id: number;
  companyId: number;
  fullName: string;
  phone: string;
  position: string;
  salaryType: EmployeeSalaryType;
  hourlyRate?: number | null;
  monthlySalary?: number | null;
  qrCode?: string | null;
  telegramId?: string | null;
  createdAt: string;
}

export type CreateEmployeeRequestSalaryType =
  (typeof CreateEmployeeRequestSalaryType)[keyof typeof CreateEmployeeRequestSalaryType];

export const CreateEmployeeRequestSalaryType = {
  hourly: "hourly",
  monthly: "monthly",
} as const;

export interface CreateEmployeeRequest {
  fullName: string;
  phone: string;
  position: string;
  salaryType: CreateEmployeeRequestSalaryType;
  hourlyRate?: number;
  monthlySalary?: number;
  telegramId?: string;
}

export type UpdateEmployeeRequestSalaryType =
  (typeof UpdateEmployeeRequestSalaryType)[keyof typeof UpdateEmployeeRequestSalaryType];

export const UpdateEmployeeRequestSalaryType = {
  hourly: "hourly",
  monthly: "monthly",
} as const;

export interface UpdateEmployeeRequest {
  fullName?: string;
  phone?: string;
  position?: string;
  salaryType?: UpdateEmployeeRequestSalaryType;
  hourlyRate?: number;
  monthlySalary?: number;
  telegramId?: string;
}

export interface EmployeeListResponse {
  data: Employee[];
  total: number;
  page: number;
  limit: number;
}

export interface QrCodeResponse {
  qrCode: string;
  employeeId: number;
}

export interface ScanRequest {
  qrData: string;
  deviceId: number;
  photo?: string | null;
}

export type ScanResponseAction =
  (typeof ScanResponseAction)[keyof typeof ScanResponseAction];

export const ScanResponseAction = {
  check_in: "check_in",
  check_out: "check_out",
} as const;

export interface Attendance {
  id: number;
  employeeId: number;
  companyId: number;
  checkIn?: string | null;
  checkOut?: string | null;
  workHours?: number | null;
  deviceId?: number | null;
  createdAt: string;
  employee: Employee;
}

export interface ScanResponse {
  action: ScanResponseAction;
  employee: Employee;
  attendance: Attendance;
  message: string;
}

export interface AttendanceListResponse {
  data: Attendance[];
  total: number;
  page: number;
  limit: number;
}

export interface Device {
  id: number;
  companyId: number;
  deviceName: string;
  location: string;
  createdAt: string;
}

export interface CreateDeviceRequest {
  deviceName: string;
  location: string;
}

export interface DeviceListResponse {
  data: Device[];
  total: number;
}

export interface PayrollRecord {
  id: number;
  employeeId: number;
  companyId: number;
  month: number;
  year: number;
  totalHours: number;
  totalDays: number;
  grossSalary: number;
  employee: Employee;
  createdAt: string;
}

export interface PayrollListResponse {
  data: PayrollRecord[];
  total: number;
  totalAmount: number;
}

export interface CalculatePayrollRequest {
  month: number;
  year: number;
}

export interface EmployeeAttendanceSummary {
  employee: Employee;
  presentDays: number;
  absentDays: number;
  lateDays: number;
  totalHours: number;
  avgArrivalTime?: string | null;
}

export interface AttendanceSummaryResponse {
  month: number;
  year: number;
  employeeSummaries: EmployeeAttendanceSummary[];
}

export type ListEmployeesParams = {
  search?: string;
  position?: string;
  page?: number;
  limit?: number;
};

export type ListAttendanceParams = {
  employee_id?: number;
  date_from?: string;
  date_to?: string;
  page?: number;
  limit?: number;
};

export type ListPayrollParams = {
  month?: number;
  year?: number;
  employee_id?: number;
};

export type GetAttendanceSummaryParams = {
  month?: number;
  year?: number;
};