# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HR Workforce Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Company admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Company registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /attendance/scan:
    post:
      operationId: scanAttendance
      tags: [attendance]
      summary: Process QR scan for attendance (check-in or check-out)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanRequest"
      responses:
        "200":
          description: Attendance recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResponse"
        "400":
          description: Invalid QR or error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: employee_id
          in: query
          schema:
            type: integer
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceListResponse"
...
  /devices:
    get:
      operationId: listDevices
      tags: [devices]
      summary: List scanner devices
      responses:
        "200":
          description: List of devices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceListResponse"
    post:
      operationId: createDevice
      tags: [devices]
      summary: Register a new scanner device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeviceRequest"
      responses:
        "201":
          description: Device created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /devices/{id}:
    put:
      operationId: updateDevice
      tags: [devices]
      summary: Update device
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeviceRequest"
      responses:
        "200":
          description: Device updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
    delete:
      operationId: deleteDevice
      tags: [devices]
      summary: Delete device
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    RegisterRequest:
      type: object
      properties:
        companyName:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        adminName:
          type: string
      required:
        - companyName
        - phone
        - email
        - password
        - adminName
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: "#/components/schemas/AdminUser"
        company:
          $ref: "#/components/schemas/Company"
      required:
        - success
        - user
        - company
...
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
      required:
        - id
        - companyId
        - name
        - email
        - role
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        logo:
          type: string
          nullable: true
        subscriptionPlan:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - phone
        - email
        - subscriptionPlan
        - createdAt
...
    Employee:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        position:
          type: string
        salaryType:
          type: string
          enum: [hourly, monthly]
        hourlyRate:
          type: number
          nullable: true
        monthlySalary:
          type: number
          nullable: true
        qrCode:
          type: string
          nullable: true
        telegramId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - companyId
        - fullName
        - phone
        - position
        - salaryType
        - createdAt
...
    ScanRequest:
      type: object
      properties:
        qrData:
          type: string
        deviceId:
          type: integer
        photo:
          type: string
          nullable: true
      required:
        - qrData
        - deviceId
    ScanResponse:
      type: object
      properties:
        action:
          type: string
          enum: [check_in, check_out]
        employee:
          $ref: "#/components/schemas/Employee"
        attendance:
          $ref: "#/components/schemas/Attendance"
        message:
          type: string
      required:
        - action
        - employee
        - attendance
        - message
    Attendance:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        companyId:
          type: integer
        checkIn:
          type: string
          format: date-time
          nullable: true
        checkOut:
          type: string
          format: date-time
          nullable: true
        workHours:
          type: number
          nullable: true
        deviceId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        employee:
          $ref: "#/components/schemas/Employee"
      required:
        - id
        - employeeId
        - companyId
        - createdAt
        - employee
    AttendanceListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Attendance"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    Device:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        deviceName:
          type: string
        location:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - companyId
        - deviceName
        - location
        - createdAt
    CreateDeviceRequest:
      type: object
      properties:
        deviceName:
          type: string
        location:
          type: string
      required:
        - deviceName
        - location
    DeviceListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Device"
        total:
          type: integer
      required:
        - data
        - total
...