# lib/api-zod/src/generated/types/attendanceListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HR Workforce Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Attendance } from "./attendance";

export interface AttendanceListResponse {
  data: Attendance[];
  total: number;
  page: number;
  limit: number;
}