# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HR Workforce Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Company {
  id: number;
  name: string;
  phone: string;
  email: string;
  logo?: string | null;
  subscriptionPlan: string;
  createdAt: Date;
}