# lib/api-zod/src/generated/types/createDeviceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HR Workforce Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDeviceRequest {
  deviceName: string;
  location: string;
}