# lib/api-zod/src/generated/types/deviceListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HR Workforce Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Device } from "./device";

export interface DeviceListResponse {
  data: Device[];
  total: number;
}