# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HR Workforce Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeSalaryType } from "./employeeSalaryType";

export interface Employee {
  id: number;
  companyId: number;
  fullName: string;
  phone: string;
  position: string;
  salaryType: EmployeeSalaryType;
  hourlyRate?: number | null;
  monthlySalary?: number | null;
  qrCode?: string | null;
  telegramId?: string | null;
  createdAt: Date;
}